Update API OAuth
