+++ package.json
@@ -10,8 +10,8 @@
   },
   "dependencies": {
     "@auth0/nextjs-auth0": "^2.3.1",
-    "amqplib": "^0.10.3",
     "@headlessui/react": "^1.7.14",
+    "amqplib": "^0.10.3",
     "animejs": "^3.2.1",
     "axios": "^1.4.0",
     "bad-words": "^3.0.4",
